latest

+++ migrations/2023-05-30-184445_create-points/up.sql
@@ -2,6 +2,6 @@
 CREATE TABLE balance(
   `id` INTEGER AUTO_INCREMENT,
   `user_id` VARCHAR(191) NOT NULL,
-  `points` INTEGER NOT NULL,
+  `points` FLOAT NOT NULL,
   primary key(`id`)
 );
